Separate e2e playwright tests from vitest unit tests.
